More Improvements to Endian Bitness for sasindex
Additional SASIndex patches that apply on top of the previous patch below.

I found a number of bugs that impact large indexes (129+) where node split and
0/NULL keys where confusing SASIndexEnum. Added a new test case to sasindex_t.c
to expose/debug this. With these patches it now now handles indexes with
1000000+ keys.

Added a new test (sasindex_tt.c) to measure performance and allow
profiling/tuning of Index performance.

I changed ./src/Makefile.am so will need to autoreconf and commit the updatedi
configure files to github.

Should look at the SASStringBTree implementation to see if similar bugs are
there. Note that StringBTree does not currently handle null strings as keys,
so may not be as large and issue.