add multi-producer-multi-consumer queues
A new set of APIs is added, to support multi-producer-multi-consumer
queues.  The API is based largely on the existing single-producer-
single-consumer API.

Of note, only the "Direct" style interface is supported.

Also of note is that the only synchronization method supported is
transactional memory (TM).  With the "--enable-htm" configure option
and on systems that support hardware transactional memory (HTM), HTM
is used (this is currently enabled only for systems which support
"cpu=power8").  All other systems will use the GNU implementation
of TM, and GNU TM is also used as a fallback on power8 when
persistent failures are reported.

Signed-off-by:  Paul A. Clarke <pc@us.ibm.com>