[MPMC] multi-producer multi-consumer queue implementation
New implementation of multi-producer multi-consumer queues.

The API is heavily based on SPHDE's single-producer single-consumer
queue API.  The persistent header block is slightly different,
necessitation a new block identifier, due to performance
improvement seen when the queue's head and tail pointers are
on different cachelines.

The implementation uses Hardware Transactional Memory, if
specifically enabled with "./configure --enable-htm".
Default (and fallback) uses GNU Transactional Memory.
This makes the implementation fairly easy and performance
is not bad.

The "Spin" and "SpinPause" API calls are not included in
this set of patches because we have not yet settled on a
final method of backing off and letting peers proceed, which
remains a work in progress.

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>