Get rid of warnings when building with GCC 6
Fix the indentation on src/ultree.cpp that was causing a warning on GCC
6.  Remove some trailing white-spaces as well.

On src/ultree.cpp, "this" has attribute nonnull, so it isn't necessary
to test if it's really not null.  Checking it can cause warnings on GCC 6.

Remove unused variables and define debugging variables only when
necessary on src/tests/sphlflogger_ttt.c.

Signed-off-by: Tulio Magno Quites Machado Filho <tuliom@linux.vnet.ibm.com>