[MPMCQ] remove use of GNU Transactional Memory
Use of GNU Transactional Memory, because of its need to support
Software Transactional Memory (STM), requires wrapping every access
writes as well as reads, to possibly contended memory inside
"__transaction_atomic".

The lock-free structures within SPHDE use tiny inline functions
to mark queue entries "complete". It is not desirable nor performant
to wrap these functions (which are realized as a single store
instruction) within transactions, nor is it (currently) desirable
to create a new "GNU-TM" version of those functions, as that would
force the user of the API to know whether they are using GNU-TM or
not. An API which can transparently use HTM for performance and
GNU-TM as a fallback is preferred, but not easily implemented
given the above requirements.

For now, remove use of GNU-TM completely, leaving the MPMC queue
implementation with HTM support only, which implies support of
the MPMC queue API only on platforms that support HTM.

The MPMC test case will not be built when the compiler does not
support HTM, and will run, but only to report success, on systems
that do not support HTM.

Note that the API has changed for affected functions. Function names changed s/TM/HTM/:

    SPHMPMCQAllocStrideDirectHTM
    SPHMPMCQGetNextCompleteDirectHTM

Signed-off-by: Paul A. Clarke pc@us.ibm.com