Fix "'this' pointer null" warning for GCC11.
The initRegion function of sassim.cpp initializes a number of
uLongTreeNode lists. Normally uLongTreeNode lists can be empty
(root pointer NULL) at any time. This is handled internally
(correctly) even in the case of NULL "this" (insert does not depend
on "this" and uses the explicite "**root" parameter.

For SPHDE initRegion is the only place where the compiler can use
static analisis to detect "'this' pointer null". This is persistent
storage so normal operator new and constructor rules do not apply.
So we expose the internals of uLongTreeNode to sassim and provide a
static inline function to: SAS allocate and initialize a
uLongTreeNode with key/value. and store that point in the root.

	* src/sassim.cpp (initULTreeList): New function.
	(initRegion): replace insertNode call with initULTreeList,
	for lists uncommitted and region.

Signed-off-by: Steven Munroe <munroesj52@gmail.com>