Allow the Used block List tree to combine adjacent blocks.
The current implementation of the used block tree is effectively a
linear list of allocated block ordered by size then address.
This creates two problems:
1) It can get very large and may cause the anchor block heap to run
out of space.
2) The list becomes a singly linked list (not a tree) so searches,
inserts and removes slow down.

The change compacts the used block tree on block allocation by:
combining adjacent even/odd blocks of the same size into a single
block of the next larger power of 2 size. This is applied recursively
if larger blocks can be combined.

On deallocation the Used List is first searched for a node with match
size and address. If found, this node is removed. Otherwise search
(p2FindCont) for a block, of larger size, containing the deallocated
block. If found, this block is split (p2split). The half that does not
contain the deallocated block, is reinserted in the list. The
containing half is checked for exact size and address match. If size
and address match, we are done and can insert the deallocated block
into the free list. Otherwise we continue spliting and containment
checking checking until we get down to the correct size and address.

	* src/sasanchr.h (regionFlags): Define bit fields for
	configuration flags.
	(rFlags): Replaces reserved1 field in SASAnchor_t struct.

	* src/sassim.cpp: add include <semaphore.h>.
	(SASS2Log2): New function.
	(p2FindCont, p2split, p2AddUsedCompact):
	New internal functions.
	(p2RemUsed [__SASDebugPrint__]): Additional debug printfs.
	(p2RemUsed): Add code the detect blocks contained within
	combined blocks and slit them as needed to isolate the
	orginal block for removal.
	(p2Dealloc [__SASDebugPrint__]): Additional debug printfs.
	(getSASUseListFlag, setSASLinearUseList, setSASCompactUseList):
	New function.
	(SASBlockAllocNoLock): If rFlags.compactUseList call
	p2AddUsedCompact else p2AddUsed.
	(SASJoinRegionByName): If creating and initializing a new
	SAS-store, set default using setSASCompactUseList.
	(SASMaxDepthAllocatedRegion, SASMaxDepthUseMem) New Functions.

	* src/sassim.h (SASS2Log2, getSASUseListFlag,
	setSASLinearUseList, setSASCompactUseList,
	SASMaxDepthAllocatedRegion, SASMaxDepthUseMem): New externs

	* src/sasutil.c (sasutil_stat_cmd): New local vars;
	cUsed, mUsed, cUsedReg, mUsedReg. Display Use List flag and
	node count and max depth for Use-List and Allocated-Region.
	(sasutil_detail_cmd): Display Use List flag,
	node count and max depth for Use-List and Allocated-Region.

	* src/tests/sassim_t.c [ARRAY_SIZE, MAX_ADDR_LIST,
	MAX_LINE_LEN]: New macros.
	(sassim_display_uselist_flag, sassim_display_summary,
	sassim_display_detail): New functions.
	[TEST_BLOCKS]: New macros.
	(blocks_s, blocks_m, blocks_l): New static arrays.
	(sassim_alloc_uselist): New unit test function.
	(main): Call sassim_alloc_uselist.

Signed-off-by: Steven Munroe <munroesj52@gmail.com>