Fixes for build test failure #61.
As SPHDE builds with fortify (__FORTIFY_SOURCE), test cases using
SASAllocateShmID may fail because a const is used as the shared
memory address passed to SASAllocateShmID then passed to memset.
So create a second version SASAllocateShmID_clear where memset
is called after successful shmat and using the returned pointer.
Use the new SASAllocateShmID_clear for the failure cases and remove
the memset call.

	* src/sasshm.c (SASAllocateShmID_clear): New function.
	* src/sasshm.h (SASAllocateShmID_clear): New extern.

	* src/tests/sascompoundheap_t.c
	(sassim_compound_heap_test1): Replace SASAllocateShmID call
	with SASAllocateShmID_clear.
	* src/tests/sassim_t.c (sassim_heap_test2):
	Replace SASAllocateShmID call with SASAllocateShmID_clear.

Signed-off-by: Steven Munroe <munroesj52@gmail.com>