Fix failures on sphgtod_t and sphgettime_t, Issue #56
This can only happen on PPC as sphde was trying meet hard-realtime
requirement by using the hardware timebase directly and avoiding the
gettimeofday() and clock_gettime() syscalls.

This can experience "lost-time" in highly virtualized environments
where the virtual machine will stop the user-level timebase, while
switched to another VM. but the clock on the wall continues.

So change the error return code when ALLOWABLE_DRIFT is exceeded
to be 77 (skip test).

 	* src/tests/sphgettime_t.c (main):
 	Change error return code to 77.
 	* src/tests/sphgtod_t.c (main):
 	Change error return codes to 77.

Signed-off-by: Steven Munroe <munroesj52@gmail.com>